requirements.txt
asgiref==3.3.1
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255081583/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1607635341180/work
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
Markdown==3.3.3
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytz==2020.5
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml==0.15.87
six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse==0.4.1
tqdm @ file:///tmp/build/80754af9/tqdm_1608922753628/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
